Move MarkdownRenderable tests
